\name{abstract}
\alias{abstract}

\title{  Parse abstract from pmc XML  }
\description{

}
\usage{
abstract(doc, exdent = 0, ...)
}

\arguments{
  \item{doc}{

}
  \item{exdent}{

}
  \item{\dots}{

}
}
\details{

}
\value{





}
\references{

}
\author{

}
\note{

}



\seealso{

}
\examples{

}

\keyword{ ~kwd1 }
