\name{addAttr}
\alias{addAttr}

\title{ 
 Copy attributes to data.frame

}
\description{
copies extra attributes from data.frame 1 to data.frame 2 , which are lost after subsetting
}
\usage{
addAttr(x, y)
}

\arguments{
  \item{x}{

}
  \item{y}{

}
}
\details{

}
\value{





}
\references{

}
\author{

}
\note{

}



\seealso{

}
\examples{

}

\keyword{ ~kwd1 }

