\name{is.xml}
\alias{is.xml}

\title{

}
\description{

}
\usage{
is.xml(doc)
}

\arguments{
  \item{doc}{

}
}
\details{

}
\value{





}
\references{

}
\author{

}
\note{

}



\seealso{

}
\examples{





function (doc) 
{
    ifelse(class(doc)[1] == "XMLInternalDocument", TRUE, FALSE)
  }
}


\keyword{ ~kwd1 }

