\name{pmcXML-internal}
\alias{fixTypes}
\alias{parse_pmc_XML}
\alias{seq2}
\alias{removeSpecChar}
\alias{htmlParse2}
\alias{xvalues}
\alias{fixText}
\alias{matchTag}

\title{Internal pmcXML functions}
\description{
  Internal package functions.
}
\note{
  Some internal functions.
}
\keyword{internal}


