\name{genomes-internal}
\alias{addAttr}
\alias{fixTypes}
\alias{matchTag}
\alias{parse_pmc_XML}
\alias{parse_pubmed_XML}
\alias{seq2}
\alias{pmid}
\alias{}
\alias{}
\alias{}


\title{Internal pubmed functions}
\description{
  Internal pubmed functions.
}
\note{
  Some internal functions.
}
\keyword{internal}
