\name{read.xls.all}
\alias{read.xls.all}

\title{ Read all sheets in  Excel files }
\description{ Read Microsoft Excel sheets into a list of data frames  }

\usage{
read.xls.all(file, ...)
}

\arguments{
  \item{file}{ path to file }
  \item{\dots}{ other options passed to \code{\link{read.xls2}} }
}
%\details{   }
\value{ A list of data frames }

% \references{ }
\author{ Chris Stubben}
% \note{ }
%\seealso{ }
% \examples{ }
\keyword{ methods }

