\name{readMarkdown}
\alias{readMarkdown}
\title{ Read markdown file }
\description{ Read markdown file into a list of subsections }
\usage{
readMarkdown(file, sentence = TRUE) }
\arguments{
  \item{file}{ Markdown file }
  \item{sentence}{ split paragraphs into sentences}
}
%\details{ }
\value{ a list of subsections }
%\references{ }
\author{ Chris Stubben }
%\note{ }
%\seealso{ }
%\examples{ }
\keyword{ methods }
