\name{searchTable}
\alias{searchTable}

\title{

}
\description{

}
\usage{
searchTable(x, prefix)
}

\arguments{
  \item{x}{

}
  \item{prefix}{

}
}
\details{

}
\value{





}
\references{

}
\author{

}
\note{

}



\seealso{

}
\examples{





function (x, prefix) 
{
    any(grepl(prefix, x, ignore.case = TRUE))
  }
}


\keyword{ ~kwd1 }

