\name{pdf2df}
\alias{pdf2df}

\title{ test

}
\description{

}
\usage{
pdf2df(x, split, captionRow = 1, headerRow = 2, labels)
}

\arguments{
  \item{x}{

}
  \item{split}{

}
  \item{captionRow}{

}
  \item{headerRow}{

}
  \item{labels}{

}
}
\details{

}
\value{


}
\references{

}
\author{

}
\note{

}

\seealso{

}
\examples{

}


\keyword{ ~kwd1 }

