\name{genomes-internal}
\alias{addAttr}
\alias{fixTypes}
\alias{matchTag}
\alias{parse_pmc_XML}
\alias{parse_pubmed_XML}
\alias{pmid}
\alias{searchTable}
\alias{seq2}
\alias{sr}
\alias{xvalues}


\title{Internal pubmed functions}
\description{
  Internal pubmed functions.
}
\note{
  Some internal functions.
}
\keyword{internal}
