\name{readMarkdown}
\alias{readMarkdown}
\title{ Read markdown file }
\description{ Read markdown file into a list of subsections }
\usage{
readMarkdown(file, sentence = TRUE, openNLP = FALSE) }
\arguments{
  \item{file}{ Markdown file }
  \item{sentence}{ split paragraphs into sentences}
  \item{openNLP}{ use sentDetect in openNLP }

}
\details{
}
\value{
}
\references{
}
\author{
}
\note{
}


\seealso{
}
\examples{


}
\keyword{ methods }
