\name{unabbrev}
\alias{unabbrev}

\title{ test

}
\description{

}
\usage{
unabbrev(species)
}

\arguments{
  \item{species}{

}
}
\details{

}
\value{





}
\references{

}
\author{

}
\note{

}



\seealso{

}
\examples{

 y<-unabbrev(c("Y. pestis", "Y. pseduotb", "Yersinia pestis", "Z. formosa" ))
y


\keyword{ methods }

